+++ backend/de.metas.business.rest-api-impl/src/main/java/de/metas/rest_api/v1/shipping/mpackage/ShippingPackageService.java
@@ -30,8 +30,8 @@ import com.google.common.collect.ImmutableSet;
 import de.metas.common.shipping.v1.shipment.JsonPackage;
 import de.metas.common.shipping.v1.shipment.mpackage.JsonCreateShippingPackageInfo;
 import de.metas.common.shipping.v1.shipment.mpackage.JsonCreateShippingPackagesRequest;
-import de.metas.handlingunits.IHUShipperTransportationBL;
-import de.metas.handlingunits.impl.AddTrackingInfosForInOutWithoutHUReq;
+import de.metas.handlingunits.shipping.IHUShipperTransportationBL;
+import de.metas.handlingunits.shipping.AddTrackingInfosForInOutWithoutHUReq;
 import de.metas.handlingunits.shipmentschedule.spi.impl.PackageInfo;
 import de.metas.inout.InOutId;
 import de.metas.inout.ShipmentScheduleId;
