+++ backend/de.metas.handlingunits.base/src/main/java/de/metas/handlingunits/shipping/AddTrackingInfosForInOutWithoutHUReq.java
@@ -20,7 +20,7 @@
  * #L%
  */
 
-package de.metas.handlingunits.impl;
+package de.metas.handlingunits.shipping;
 
 import de.metas.handlingunits.shipmentschedule.spi.impl.PackageInfo;
 import de.metas.inout.InOutId;
