+++ backend/de.metas.handlingunits.base/src/main/java/de/metas/handlingunits/shipping/IHUShipperTransportationBL.java
@@ -1,8 +1,7 @@
-package de.metas.handlingunits;
+package de.metas.handlingunits.shipping;
 
 import com.google.common.collect.ImmutableList;
-import de.metas.handlingunits.impl.AddTrackingInfosForInOutWithoutHUReq;
-import de.metas.handlingunits.shipping.CreatePackagesForInOutRequest;
+import de.metas.handlingunits.IHUQueryBuilder;
 import de.metas.handlingunits.model.I_M_HU;
 import de.metas.shipping.model.I_M_ShipperTransportation;
 import de.metas.shipping.model.I_M_ShippingPackage;
