+++ backend/de.metas.handlingunits.base/src/main/java/de/metas/handlingunits/shipping/impl/HUPackageBL.java
@@ -3,7 +3,7 @@ package de.metas.handlingunits.shipping.impl;
 import com.google.common.collect.ImmutableSet;
 import de.metas.handlingunits.shipping.IHUPackageBL;
 import de.metas.handlingunits.IHUPackageDAO;
-import de.metas.handlingunits.IHUShipperTransportationBL;
+import de.metas.handlingunits.shipping.IHUShipperTransportationBL;
 import de.metas.handlingunits.exceptions.HUException;
 import de.metas.handlingunits.model.I_M_HU;
 import de.metas.handlingunits.model.I_M_Package_HU;
